refactor: useCardNumber 안의 주요 로직 함수 및 상수들을 분리
